+++ pkg/service/invoice/create_invoice.go
@@ -36,6 +36,7 @@ func (c CreateInvoice) Call(approver models.OfficeUser, invoice *models.Invoice,
 	return c.DB.ValidateAndCreate(invoice)
 }
 
+// createInvoiceNumber creates a new invoice number for a given shipment.
 func (c CreateInvoice) createInvoiceNumber(shipment models.Shipment) (string, error) {
 	// If we have existing invoices, then get the existing base invoice number and add the appropriate suffix,
 	// then go ahead and return it.
@@ -69,10 +70,43 @@ func (c CreateInvoice) createInvoiceNumber(shipment models.Shipment) (string, er
 	}
 	year := shipment.CreatedAt.In(loc).Year()
 
-	invoiceNumber, err := models.GenerateBaseInvoiceNumber(c.DB, scac, year)
+	invoiceNumber, err := generateBaseInvoiceNumber(c.DB, scac, year)
 	if err != nil {
 		return "", errors.Wrap(err, "Could not generate invoice number")
 	}
 
 	return invoiceNumber, nil
+}
+
+// generateBaseInvoiceNumber creates a new base invoice number (the first for a shipment) for a given SCAC/year.
+func generateBaseInvoiceNumber(db *pop.Connection, scac string, year int) (string, error) {
+	if len(scac) == 0 {
+		return "", errors.New("SCAC cannot be nil or empty string")
+	}
+
+	if year <= 0 {
+		return "", errors.Errorf("Year (%d) must be non-negative", year)
+	}
+
+	var sequenceNumber int
+	sql := `INSERT INTO invoice_number_trackers as trackers (standard_carrier_alpha_code, year, sequence_number)
+			VALUES ($1, $2, 1)
+		ON CONFLICT (standard_carrier_alpha_code, year)
+		DO
+			UPDATE
+				SET sequence_number = trackers.sequence_number + 1
+				WHERE trackers.standard_carrier_alpha_code = $1 AND trackers.year = $2
+		RETURNING sequence_number
+	`
+
+	err := db.RawQuery(sql, scac, year).First(&sequenceNumber)
+	if err != nil {
+		return "", errors.Wrapf(err, "Error when incrementing invoice sequence number for %s/%d", scac, year)
+	}
+
+	if sequenceNumber > 9999 {
+		return "", errors.Errorf("All four-digit invoice sequence numbers already used for %s/%d", scac, year)
+	}
+
+	return fmt.Sprintf("%s%d%04d", scac, year%100, sequenceNumber), nil
 }
